+++ CHANGELOG.md
@@ -3,6 +3,9 @@
 *   New Features
     *   Add an option to sort Up Next by date
         ([#4586](https://github.com/Automattic/pocket-casts-android/pull/4586))
+*   Updates
+    *   Improve sleep timer terminology and chapter option visibility
+        ([#4564](https://github.com/Automattic/pocket-casts-android/pull/4564))
 *   Bug Fixes
     *   Fix restoring large Up Next queues
         ([#4562](https://github.com/Automattic/pocket-casts-android/pull/4562))